settings: Move ParserChoice into types module
In preparation for storing parser choice inside DepsSource.
